move magnet out, add docstrings for refl helpers
