+++ src/ibex_bluesky_core/plans/__init__.py
@@ -1,6 +1,13 @@
 from bluesky import plan_stubs as bps
 
+from ibex_bluesky_core.devices import get_pv_prefix
 from ibex_bluesky_core.devices.simpledae import SimpleDae
+from ibex_bluesky_core.devices.simpledae.controllers import (
+    PeriodPerPointController,
+    RunPerPointController,
+)
+from ibex_bluesky_core.devices.simpledae.reducers import MonitorNormalizer
+from ibex_bluesky_core.devices.simpledae.waiters import GoodFramesWaiter, PeriodGoodFramesWaiter
 
 
 def set_num_periods(dae: SimpleDae, nperiods: int):
@@ -10,3 +17,42 @@ def set_num_periods(dae: SimpleDae, nperiods: int):
         raise ValueError(
             f"Could not set {nperiods} periods on DAE (probably requesting too many points, or already running)"
         )
+
+
+DEFAULT_MON = 1
+
+
+def common_dae(
+    *,
+    det_pixels: list[int],
+    frames: int,
+    periods: bool = True,
+    monitor: int = DEFAULT_MON,
+    save_run: bool = False,
+):
+    """TODO explain this is really just a shortcut for doing the below"""
+    prefix = get_pv_prefix()
+
+    if periods:
+        controller = PeriodPerPointController(save_run=save_run)
+        waiter = PeriodGoodFramesWaiter(frames)
+    else:
+        controller = RunPerPointController(save_run=save_run)
+        waiter = GoodFramesWaiter(frames)
+
+    reducer = MonitorNormalizer(
+        prefix=prefix,
+        detector_spectra=det_pixels,
+        monitor_spectra=[monitor],
+    )
+
+    dae = SimpleDae(
+        prefix=prefix,
+        controller=controller,
+        waiter=waiter,
+        reducer=reducer,
+    )
+
+    dae.reducer.intensity.set_name("intensity")  # type: ignore
+    dae.reducer.intensity_stddev.set_name("intensity_stddev")  # type: ignore
+    return dae
