+++ src/ibex_bluesky_core/plans/muons/hifi_magnet_scan.py
@@ -2,7 +2,6 @@
 RE(magnet_scan("X", 0, 1, 2)).
 """
 
-import asyncio
 from pathlib import Path
 
 import bluesky.plans as bp
@@ -10,13 +9,7 @@ import bluesky.preprocessors as bpp
 import matplotlib
 import matplotlib.pyplot as plt
 from bluesky.callbacks import LiveFitPlot, LiveTable
-from ophyd_async.core import (
-    AsyncStatus,
-    HintedSignal,
-    StandardReadable,
-    observe_value,
-)
-from ophyd_async.epics.signal import epics_signal_r, epics_signal_rw, epics_signal_w
+from ophyd_async.epics.signal import epics_signal_r
 from ophyd_async.plan_stubs import ensure_connected
 
 from ibex_bluesky_core.callbacks.file_logger import HumanReadableFileCallback
@@ -26,6 +19,7 @@ from ibex_bluesky_core.callbacks.fitting.livefit_logger import LiveFitLogger
 from ibex_bluesky_core.callbacks.plotting import LivePlot
 from ibex_bluesky_core.devices import get_pv_prefix
 from ibex_bluesky_core.devices.block import block_r
+from ibex_bluesky_core.devices.muons.hifi_magnet import HIFIMagnetAxis
 from ibex_bluesky_core.run_engine import get_run_engine
 
 matplotlib.rcParams["figure.autolayout"] = True
@@ -36,42 +30,9 @@ RE = get_run_engine()
 READABLE_FILE_OUTPUT_DIR = Path("C:\\") / "instrument" / "var" / "logs" / "bluesky" / "output_files"
 
 
-class MagnetAxis(StandardReadable):
-    def __init__(self, prefix: str, axis: str) -> None:
-        with self.add_children_as_readables(HintedSignal):
-            self.setpoint = epics_signal_rw(float, f"{prefix}CS:SB:Field_{axis}_Target")
-
-        self.readback = epics_signal_r(float, f"{prefix}CS:SB:Field_{axis}")
-        self.ready = epics_signal_r(bool, f"{prefix}CS:SB:Field_{axis}_Ready")
-        self.go = epics_signal_w(bool, f"{prefix}CS:SB:Field_{axis}_Go")
-
-        super().__init__(name=f"Field_{axis}_magnet")
-        self.setpoint.set_name(f"Field_{axis}_magnet")
-
-    @AsyncStatus.wrap
-    async def trigger(self) -> None:
-        pass
-
-    @AsyncStatus.wrap
-    async def set(self, value: float) -> None:
-        """Set the setpoint."""
-        await self.setpoint.set(value, wait=True, timeout=None)
-        await asyncio.sleep(5.0)  # race conditions?
-        await self.go.set(True, wait=True, timeout=None)
-        await asyncio.sleep(5.0)
-        async for stat in observe_value(self.ready):
-            if stat:
-                break
-        await asyncio.sleep(5.0)  # ensure latest readings have been taken
-
-    def __repr__(self) -> str:
-        """Debug representation."""
-        return f"{self.__class__.__name__}(name={self.name})"
-
-
-def magnet_axis(axis: str) -> MagnetAxis:
+def magnet_axis(axis: str) -> HIFIMagnetAxis:
     prefix = get_pv_prefix()
-    return MagnetAxis(prefix=prefix, axis=axis)
+    return HIFIMagnetAxis(prefix=prefix, axis=axis)
 
 
 def magnet_scan(
@@ -82,7 +43,6 @@ def magnet_scan(
     *,
     model: Fit = Linear(),
     rel: bool = False,
-    read_block: str | None = None,
 ):
     magnet = magnet_axis(axis)
     magnetometer_x1 = epics_signal_r(float, "IN:HIFI:G3HALLPR_01:0:FIELD", name="x1")
