+++ src/ibex_bluesky_core/plans/reflectometry/crisp.py
@@ -12,16 +12,9 @@ from ophyd_async.plan_stubs import ensure_connected
 from ibex_bluesky_core.callbacks import ISISCallbacks
 from ibex_bluesky_core.callbacks.fitting import FitMethod
 from ibex_bluesky_core.callbacks.fitting.fitting_utils import Fit, Linear
-from ibex_bluesky_core.devices import get_pv_prefix
 from ibex_bluesky_core.devices.simpledae import SimpleDae
-from ibex_bluesky_core.devices.simpledae.controllers import (
-    PeriodPerPointController,
-    RunPerPointController,
-)
-from ibex_bluesky_core.devices.simpledae.reducers import MonitorNormalizer
-from ibex_bluesky_core.devices.simpledae.waiters import GoodFramesWaiter, PeriodGoodFramesWaiter
 from ibex_bluesky_core.plan_stubs import call_qt_aware
-from ibex_bluesky_core.plans import set_num_periods
+from ibex_bluesky_core.plans import common_dae, set_num_periods
 from ibex_bluesky_core.plans.reflectometry import centred_pixel, refl_parameter
 from ibex_bluesky_core.run_engine import get_run_engine
 
@@ -44,31 +37,9 @@ def crisp_dae(
     monitor: int = 1,
     save_run: bool = False,
 ) -> SimpleDae:
-    prefix = get_pv_prefix()
-
-    if periods:
-        controller = PeriodPerPointController(save_run=save_run)
-        waiter = PeriodGoodFramesWaiter(frames)
-    else:
-        controller = RunPerPointController(save_run=save_run)
-        waiter = GoodFramesWaiter(frames)
-
-    reducer = MonitorNormalizer(
-        prefix=prefix,
-        detector_spectra=det_pixels,
-        monitor_spectra=[monitor],
+    return common_dae(
+        det_pixels=det_pixels, frames=frames, periods=periods, monitor=monitor, save_run=save_run
     )
-
-    dae = SimpleDae(
-        prefix=prefix,
-        controller=controller,
-        waiter=waiter,
-        reducer=reducer,
-    )
-
-    dae.reducer.intensity.set_name("intensity")  # type: ignore
-    dae.reducer.intensity_stddev.set_name("intensity_stddev")  # type: ignore
-    return dae
 
 
 def scan(
